Introduce FieldCollector
