+++ src/Execution/Executor.php
@@ -7,12 +7,7 @@ use Digia\GraphQL\Error\GraphQLException;
 use Digia\GraphQL\Error\InvalidTypeException;
 use Digia\GraphQL\Error\UndefinedException;
 use Digia\GraphQL\Language\Node\FieldNode;
-use Digia\GraphQL\Language\Node\FragmentDefinitionNode;
-use Digia\GraphQL\Language\Node\FragmentSpreadNode;
-use Digia\GraphQL\Language\Node\InlineFragmentNode;
-use Digia\GraphQL\Language\Node\NodeInterface;
 use Digia\GraphQL\Language\Node\OperationDefinitionNode;
-use Digia\GraphQL\Language\Node\SelectionSetNode;
 use Digia\GraphQL\Schema\Schema;
 use Digia\GraphQL\Type\Definition\AbstractTypeInterface;
 use Digia\GraphQL\Type\Definition\Field;
@@ -29,7 +24,6 @@ use function Digia\GraphQL\Type\SchemaMetaFieldDefinition;
 use function Digia\GraphQL\Type\TypeMetaFieldDefinition;
 use function Digia\GraphQL\Type\TypeNameMetaFieldDefinition;
 use function Digia\GraphQL\Util\toString;
-use function Digia\GraphQL\Util\typeFromAST;
 
 /**
  * Class AbstractStrategy
@@ -52,6 +46,12 @@ class Executor
      */
     protected $rootValue;
 
+
+    /**
+     * @var FieldCollector
+     */
+    protected $fieldCollector;
+
     /**
      * @var array
      */
@@ -71,11 +71,13 @@ class Executor
     public function __construct(
         ExecutionContext $context,
         OperationDefinitionNode $operation,
+        FieldCollector $fieldCollector,
         $rootValue
     ) {
-        $this->context   = $context;
-        $this->operation = $operation;
-        $this->rootValue = $rootValue;
+        $this->context        = $context;
+        $this->operation      = $operation;
+        $this->fieldCollector = $fieldCollector;
+        $this->rootValue      = $rootValue;
     }
 
 
@@ -96,7 +98,7 @@ class Executor
         $fields               = [];
         $visitedFragmentNames = [];
         try {
-            $fields = $this->collectFields(
+            $fields = $this->fieldCollector->collectFields(
                 $objectType,
                 $this->operation->getSelectionSet(),
                 $fields,
@@ -153,126 +155,6 @@ class Executor
                     [$operation]
                 );
         }
-    }
-
-    /**
-     * @param ObjectType       $runtimeType
-     * @param SelectionSetNode $selectionSet
-     * @param                  $fields
-     * @param                  $visitedFragmentNames
-     * @return mixed
-     * @throws InvalidTypeException
-     * @throws \Digia\GraphQL\Error\ExecutionException
-     * @throws \Digia\GraphQL\Error\InvariantException
-     */
-    protected function collectFields(
-        ObjectType $runtimeType,
-        SelectionSetNode $selectionSet,
-        &$fields,
-        &$visitedFragmentNames
-    ) {
-        foreach ($selectionSet->getSelections() as $selection) {
-            // Check if this Node should be included first
-            if (!$this->shouldIncludeNode($selection)) {
-                continue;
-            }
-            // Collect fields
-            if ($selection instanceof FieldNode) {
-                $fieldName = $this->getFieldNameKey($selection);
-
-                if (!isset($fields[$fieldName])) {
-                    $fields[$fieldName] = [];
-                }
-
-                $fields[$fieldName][] = $selection;
-            } elseif ($selection instanceof InlineFragmentNode) {
-                if (!$this->doesFragmentConditionMatch($selection, $runtimeType)) {
-                    continue;
-                }
-
-                $this->collectFields($runtimeType, $selection->getSelectionSet(), $fields, $visitedFragmentNames);
-            } elseif ($selection instanceof FragmentSpreadNode) {
-                $fragmentName = $selection->getNameValue();
-
-                if (!empty($visitedFragmentNames[$fragmentName])) {
-                    continue;
-                }
-
-                $visitedFragmentNames[$fragmentName] = true;
-                /** @var FragmentDefinitionNode $fragment */
-                $fragment = $this->context->getFragments()[$fragmentName];
-                $this->collectFields($runtimeType, $fragment->getSelectionSet(), $fields, $visitedFragmentNames);
-            }
-        }
-
-        return $fields;
-    }
-
-
-    /**
-     * @param $node
-     * @return bool
-     * @throws InvalidTypeException
-     * @throws \Digia\GraphQL\Error\ExecutionException
-     * @throws \Digia\GraphQL\Error\InvariantException
-     */
-    private function shouldIncludeNode(NodeInterface $node): bool
-    {
-
-        $contextVariables = $this->context->getVariableValues();
-
-        $skip = coerceDirectiveValues(SkipDirective(), $node, $contextVariables);
-
-        if ($skip && $skip['if'] === true) {
-            return false;
-        }
-
-        $include = coerceDirectiveValues(IncludeDirective(), $node, $contextVariables);
-
-        if ($include && $include['if'] === false) {
-            return false;
-        }
-
-        return true;
-    }
-
-    /**
-     * @param FragmentDefinitionNode|InlineFragmentNode $fragment
-     * @param ObjectType                                $type
-     * @return bool
-     * @throws InvalidTypeException
-     */
-    private function doesFragmentConditionMatch(
-        NodeInterface $fragment,
-        ObjectType $type
-    ): bool {
-        $typeConditionNode = $fragment->getTypeCondition();
-
-        if (!$typeConditionNode) {
-            return true;
-        }
-
-        $conditionalType = typeFromAST($this->context->getSchema(), $typeConditionNode);
-
-        if ($conditionalType === $type) {
-            return true;
-        }
-
-        if ($conditionalType instanceof AbstractTypeInterface) {
-            return $this->context->getSchema()->isPossibleType($conditionalType, $type);
-        }
-
-        return false;
-    }
-
-    /**
-     * @TODO: consider to move this to FieldNode
-     * @param FieldNode $node
-     * @return string
-     */
-    private function getFieldNameKey(FieldNode $node): string
-    {
-        return $node->getAlias() ? $node->getAlias()->getValue() : $node->getNameValue();
     }
 
     /**
@@ -961,7 +843,7 @@ class Executor
             }
         }
 
-        return $this->collectAndExecuteSubFields(
+        return $this->executeSubFields(
             $returnType,
             $fieldNodes,
             $info,
@@ -1041,7 +923,7 @@ class Executor
      * @throws \Digia\GraphQL\Error\InvariantException
      * @throws \Throwable
      */
-    private function collectAndExecuteSubFields(
+    private function executeSubFields(
         ObjectType $returnType,
         $fieldNodes,
         ResolveInfo $info,
@@ -1054,7 +936,7 @@ class Executor
         foreach ($fieldNodes as $fieldNode) {
             /** @var FieldNode $fieldNode */
             if ($fieldNode->getSelectionSet() !== null) {
-                $subFields = $this->collectFields(
+                $subFields = $this->fieldCollector->collectFields(
                     $returnType,
                     $fieldNode->getSelectionSet(),
                     $subFields,
